Add AnotherWatchdog for the system
We found that there are cases the system is occupied by some threads
(most likely some userland RT thread) and no user program have chance
to run. In this case, we lose adb connection and the system is not
responding. We don't have a mechanism to deal with this until now.
Therefore we add a feeder for a kernel softdog, which can detect this.

	new file:   AnotherWatchdogService.java
	modified:   SystemServer.java

Change-Id: I7459bbc94474ca1be2ab1ef2bbd94f0cef055bf5